requirements.txt
flask
selenium
webdriver-manager
groq
python-dotenv
